cache providers used for functions
We need to be able to reuse providers which are used for function calls,
so cache them in the EvalContext.

This does require a way to close the provider however, which we can do
via the normal root module close node since it already does the same for
provisioners. Update the `CloseProvisioners` method to be a more general
`ClosePlugins` method, and use that to call Close on the cached
providers too.